requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
langgraph>=0.0.20
anthropic>=0.25.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
sentence-transformers>=2.2.0